ContractAI/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.0.0.post2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Storage
minio==7.1.17
aiofiles==23.2.1

# Queue
redis==5.0.1
celery==5.3.4
flower==2.0.1

# AI/ML
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
sentence-transformers==2.2.2
openai==1.3.5
anthropic==0.5.0
cohere==4.37
mistralai==0.0.7
nltk==3.8.1
scikit-learn==1.3.2
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.0.1
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.1
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
prometheus-client==0.17.1